+++ src/handlers/wallets/plain.rs
@@ -8,7 +8,7 @@ pub fn index(state: &ServerState) -> base::JsonResult {
 }
 
 #[get("/plain_wallets/<id>", format = "application/json")]
-pub fn show(state: &ServerState, id: i32) -> base::JsonResult {
+pub fn show(state: &ServerState, id: u64) -> base::JsonResult {
     base::show(state, id, |wallets| wallets.plains)
 }
 
@@ -18,11 +18,11 @@ pub fn create(state: &ServerState, wallet: PlainWallet) -> base::JsonResult {
 }
 
 #[put("/plain_wallets/<id>", format = "application/json", data = "<wallet>")]
-pub fn update(state: &ServerState, id: i32, wallet: PlainWallet) -> base::JsonResult {
+pub fn update(state: &ServerState, id: u64, wallet: PlainWallet) -> base::JsonResult {
     base::update(state, id, wallet, |wallets| wallets.plains.as_mut() )
 }
 
 #[delete("/plain_wallets/<id>", format = "application/json")]
-pub fn destroy(state: &ServerState, id: i32) -> base::JsonResult {
+pub fn destroy(state: &ServerState, id: u64) -> base::JsonResult {
     base::destroy(state, id, |wallets| wallets.plains.as_mut() )
 }
